Move include handles settings to `main`
